Rename docs to swaggerspec
