+++ .golangci.yml
@@ -1,6 +1,6 @@
 run:
   skip-dirs:
-    - docs
+    - swaggerspec
     - pkg/uiserver
   timeout: 2m
